requirements.txt
typing-extensions<4.6.0
cohere
openai
openai_whisper
pandas>=1.1.0